dashboard/requirements.txt
cufflinks
dash==1.16.0
plotly
backoff==1.10.0
requests==2.24.0
python-dotenv
pyyaml